Add files via upload
I wrote a short script that computes the Dowker complex for a directed cycle graph. It uses networkx and gudhi's simplex tree. The weights are non-negative reals. I just tried it with 24 points in the cycle graph and it computes it just fine. The idea is that to check whether a k-simplex [s0, s1, ..., sk] is in the Dowker complex at filtration level T, we need to check to see if there is a point x s.t. d(si, x) <= T for all i = 0, ..., k. 

So if we go to the submatrix of the distance matrix containing all si rows, then the maximum, M(j), of the values within column j tells us that at that the point xj associated to column j is within distance  Mj of all si (for i = 0, ..., k), i.e., d(si, xj) <= Mj for all i. Doing this for all j and then taking the j* with the smallest Mj will tell us that point xj* can see all si.